util: split interface into multiple files
Added menufuctions, utility and query python files
